+++ frontend/src/metabase/query_builder/components/view/View.jsx
@@ -24,7 +24,7 @@ import QueryModals from "../QueryModals";
 import { ViewTitleHeader, ViewSubHeader } from "./ViewHeader";
 import NewQuestionHeader from "./NewQuestionHeader";
 import ViewFooter from "./ViewFooter";
-import ViewSidebar from "./ViewSidebar";
+import ViewSidebar from "./ViewSidebar/ViewSidebar";
 import QuestionDataSelector from "./QuestionDataSelector";
 
 import ChartSettingsSidebar from "./sidebars/ChartSettingsSidebar";
