+++ frontend/src/metabase/query_builder/components/view/ViewSidebar/ViewSidebar.styled.jsx
@@ -12,7 +12,6 @@ export const ViewSidebarAside = styled.aside`
     transition: none;
   }
 
-
   ${({ left }) =>
     left &&
     css`
@@ -40,10 +39,4 @@ export const ViewSidebarContent = styled.div`
   position: absolute;
   height: 100%;
   width: 100%;
-
-  ${({ widthProp: width }) =>
-    width &&
-    css`
-      width: ${width}px;
-    `}
 `;
